src/terminal_test.rs
#[cfg(test)]
mod tests {
    use super::*;

    #[test]
    fn test_get_padding_for_author() {
        // Given an author length of 10 and sentence padding of 3
        if let Ok(padding) = get_padding_for_author(10, 3) {
            // Should use LARGE_PADDING (10) since sentence_padding > 2
            assert!(padding > 0, "Padding should be positive");
            assert!(padding < get_terminal_length().unwrap(), "Padding should be less than terminal width");
        }

        // Test with small padding case
        if let Ok(padding) = get_padding_for_author(10, 2) {
            // Should use SMALL_PADDING (5) since sentence_padding <= 2
            assert!(padding > 0, "Padding should be positive");
            assert!(padding < get_terminal_length().unwrap(), "Padding should be less than terminal width");
        }
    }

    #[test]
    fn test_get_sentences_needed() {
        if let Ok(sentences) = get_sentences_according_to_terminal_and_padding(100, 2) {
            assert!(sentences > 0, "Should return at least one sentence");
        }

        // Test with empty quote
        if let Ok(sentences) = get_sentences_according_to_terminal_and_padding(0, 2) {
            assert!(sentences > 0, "Should return at least one sentence even for empty quote");
        }
    }

    #[test]
    fn test_get_lines_of_quote() {
        let test_quote = String::from("This is a test quote that should be split into multiple lines based on terminal width");
        
        if let Ok(lines) = get_lines_of_quote_according_to_terminal_and_padding(test_quote, 2) {
            assert!(!lines.is_empty(), "Should return at least one line");
            
            // Check that no line exceeds terminal width
            if let Ok(term_width) = get_terminal_length() {
                for line in lines {
                    assert!(line.len() <= term_width - 4, "Line should not exceed terminal width minus padding");
                }
            }
        }
    }

    #[test]
    fn test_get_terminal_length() {
        if let Ok(length) = get_terminal_length() {
            assert!(length > 0, "Terminal length should be positive");
            // Most terminals are at least 80 columns wide
            assert!(length >= 80, "Terminal length should be at least 80 characters");
        }
    }
}
